+++ extension-testcontainer/src/main/java/io.camunda.zeebe.process.test.extension.testcontainer/ZeebeProcessTestExtension.java
@@ -17,6 +17,7 @@ package io.camunda.zeebe.process.test.extension.testcontainer;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.zeebe.client.ZeebeClient;
+import io.camunda.zeebe.process.test.ObjectMapperConfig;
 import io.camunda.zeebe.process.test.assertions.BpmnAssert;
 import io.camunda.zeebe.process.test.filters.RecordStream;
 import java.lang.reflect.Field;
@@ -81,11 +82,11 @@ public class ZeebeProcessTestExtension
   @Override
   public void beforeEach(final ExtensionContext extensionContext) {
     final Object engineContent = getStore(extensionContext.getParent().get()).get(KEY_ZEEBE_ENGINE);
-    final Optional<ObjectMapper> customObjectMapper = getCustomMapper(extensionContext);
+    final ObjectMapper objectMapper = getCustomMapper(extensionContext);
     final ContainerizedEngine engine = (ContainerizedEngine) engineContent;
     engine.start();
 
-    final ZeebeClient client = createClient(customObjectMapper, engine);
+    final ZeebeClient client = engine.createClient(objectMapper);
     final RecordStream recordStream = RecordStream.of(new RecordStreamSourceImpl(engine));
     BpmnAssert.initRecordStream(recordStream);
 
@@ -171,42 +172,26 @@ public class ZeebeProcessTestExtension
   }
 
   /**
-   * Get a custom object mapper from the test context
+   * Get a custom object mapper from the test context or a default one if it is not provided
    *
    * @param context jUnit5 extension context
-   * @return {@link Optional} of {@link ObjectMapper}, or Optional.empty() if no object mapper are
-   *     in the context
+   * @return the custom {@link ObjectMapper} if provided, a new one if not
    */
-  private Optional<ObjectMapper> getCustomMapper(final ExtensionContext context) {
+  private ObjectMapper getCustomMapper(final ExtensionContext context) {
     final Optional<Field> customMapperOpt =
         getField(context.getRequiredTestClass(), objectMapperInstance);
     if (!customMapperOpt.isPresent()) {
-      return Optional.empty();
+      return new ObjectMapper();
     }
     final Field customMapper = customMapperOpt.get();
     ReflectionUtils.makeAccessible(customMapper);
     try {
-      return Optional.of((ObjectMapper) customMapper.get(context.getRequiredTestInstance()));
+      final ObjectMapper customObjectMapper =
+          (ObjectMapper) customMapper.get(context.getRequiredTestInstance());
+      ObjectMapperConfig.initializeCustomMapper(customObjectMapper);
+      return customObjectMapper;
     } catch (final IllegalAccessException e) {
       throw new RuntimeException(e);
     }
-  }
-
-  /**
-   * Create a {@link ZeebeClient}. If a custom {@link ObjectMapper} is provided it is initialized
-   * into the {@link ObjectMapperConfig} and it is configured into the client
-   *
-   * @param objectMapper an {@link Optional} of {@link ObjectMapper}
-   * @param engine the used engine
-   * @return a zeebe client
-   */
-  private ZeebeClient createClient(
-      final Optional<ObjectMapper> objectMapper, final ContainerizedEngine engine) {
-    if (objectMapper.isPresent()) {
-      final ObjectMapper customObjectMapper = objectMapper.get();
-      ObjectMapperConfig.initialize(customObjectMapper);
-      return engine.createClient(customObjectMapper);
-    }
-    return engine.createClient();
   }
 }
